Split GameDatasetCollectionSchema into own folder.
